+++ WHATS_NEW.md
@@ -1,6 +1,15 @@
-### August 20 2019
+### August 21 2019
+# Timestamps for Segmentation Layers
+* Under the graph tab, the new Timestamp field allows you to view segmentations at various points in time.
+* Warning: Setting a timestamp will deselect all selected segments including those that are hidden.
+* When viewing a segmentation layer in an old state, it will be highlighted and feature a red <span style="color:red">🕘</span>.
+    * ![](https://i.imgur.com/RFupZod.png)
+    * Clicking the <span style="color:red">🕘</span> will reset the layer back to the current time.
+    * This will also deselect all selected segments of that older state.
+
 # Optional Pointer Lock
- * Undo button pops up in bottom when you delete a layer, in case this was accidental	* Enabling pointer lock allowing mousedrag and ctrl + mousedrag without running into borders.
-  * Click the '⚙' icon and uncheck 'Show cursor on mouse drag' to enable.
+* Undo button pops up in bottom when you delete a layer, in case this was accidental	* Enabling pointer lock allowing mousedrag and ctrl + mousedrag without running into borders.
+ * Click the ⚙ icon and uncheck 'Show cursor on mouse drag' to enable.
+
 # Misc
 * Hotfix for segmentation button highlighting